Replace an ad-hoc script with pure Make commands
Just inline invocations of clang tools into the Makefile, do not use
an intermediate shell script as that complicates inspection.

Also, limit the effect to *.c and *.h files in $(SRC_PATH). We can
extend the formatting later to other file types.

Now that we don't have a configurable script we have to duplicate
the fixups and checks. However, that allows us to be more precise
with checks and avoid interference between fixups and checks.
With that we no longer need the "fmt_clean" target, just a regular
"clean" will do.