Improve key kind verification and tests
Previous implementation allowed to (ab)use SecureMessage class in
sign/verify mode by not specifying one of the keys. It is not possible
now since we're using the new C API. Now we require both public and
private key to be specified at Secure Message construction.

Update the tests to verify the new requirement and add more error
checking to constructor to ensure that both keys are provided and
that they have correct kinds.