Remove "cargo deadlinks" check (#403)
I have found this tool not *that* useful in preventing mistakes in API
documentation, but it adds quite a considerable chunk to the build time
of a development environment. I don't really want developers to wait
5-15 minutes to download and build all the dependencies just to have
a check which passes 99% of time.

We may reconsider this decision later if we actually run into issues
with dead links in API docs.