Formatting: automatic formatting for ThemisPP tests
Add ThemisPP tests (and C++ tests in genenral) into "make fmt" target.
While we're here, refactor the makefile a bit to reduce variable
duplication.

As with main C++ source code, we use a slightly different style than
for the C code. I start to consider that we should be formatting C code
with this style as well...