Fix Secure Cell corruption tests on i386
It turned out that current tests do not work on i386 architecture
(32-bit x86). This is because we try allocating ~0xDEADBEEF bytes
which is not possible in 32-bit userland that typically has a limit
of 2 or 3 GB of user memory. The constant we use does not really
matter but should be distinct enough. Use something smaller instead.