Formatting: update C code format
The main change here is disabling of argument "bin packing" which is a
major source of annoying formatting. We have *a lot* of functions with
long parameter lists so these are used quite oftern. Unfortunately, bin
packing these lists leads to poor readability and extremely brittle
formatting which highly depends on the length of variables involved.
Disabling bin packing gives us consistency: a function call either fits
on a single line entirely, or it is formatted with one argument per line
which is much more readable.

Other changes include introduction of include header groups which
ensures that includes are ordered and grouped consistently. We use the
"Preserve" mode because currently clang-format fails to detect main
headers correctly. However, it should at least reorder headers which
match existing groups.

Finally we do some fine-tuning of the penalties for excess characters
and line-breaking. This gives a bit nicer line breaks as clang-format
allows minor overflows past the right column edge if avoiding them
requires to introduce an inconvenient line break near assignment or
a function call. This is the part that I dislike the most because it
feels very brittle. But well...