Formatting: JsThemis native extension
Add native extension of JsThemis to "make fmt" target group.

This requires quite a bit of effort because clang-tidy insists on having
access to *all* source code involved in complation, including all those
NodeJS headers and its modules.

First of all, we have to add some hardcoded paths to NodeJS headers.
I have not found a proper way to get these paths so we just hardcode
them and hope that this is fine.

Then we need to have <nan.h> header file for which we have to install
the nan module. We do a little hack to achieve that, pretending to
'format' the installation path for Node modules.