Silent version checks
Silence output to stderr if there is no language support installed.
Most of the version checks already redirect stderr to /dev/null but some
of them merge stdout and stderr due to historical reasons. In this case
check whether a binary is available before trying to execute it.

This makes sure that "make" output does not contain unexpected messages
like

    /bin/bash: python2 not found
    /bin/bash: php not found