Remove "tests/soter/nist-sts/assess" binary
I don't know why we keep this binary in the repository but I get
triggered every time I do "make clean" and it is removed which then
results in a dirty git index.

This file will be rebuilt if necessary for NIST STS tests so there is
no reason to keep it in Git.