Formatting: JsThemis native extension (#412)
* Formatting: JsThemis native extension

Add native extension of JsThemis to "make fmt" target group.

This requires quite a bit of effort because clang-tidy insists on having
access to *all* source code involved in complation, including all those
NodeJS headers and its modules.

First of all, we have to add some hardcoded paths to NodeJS headers.
I have not found a proper way to get these paths so we just hardcode
them and hope that this is fine.

Then we need to have <nan.h> header file for which we have to install
the nan module. We do a little hack to achieve that, pretending to
'format' the installation path for Node modules.

* Add NodeJS to format checking image

Install npm and nodejs as we need them for checking JsThemis native
source code. At least for now.

* Add a warning if NodeJS is not installed

Print out a warning when formatting files if NodeJS is not installed and
we are not able to reformat JsThemis files.

* Add symlink to jsthemis.mk for Rust's libthemis-src

The current build system for libthemis-src is quite messed up and
requires peculiar symlinking to work. Add a symlink for the new file
included from the main Makefile.