Enable more compiler warnings (#377)
* Enable more compiler warnings

We *are* security-oriented project so let's enable as many relevant
flags as possible. Here are some flags that I found useful after
looking through the latest lists for GCC 8 and Clang 7. (In addition
to catch-all "-Wall -Wextra" which enable most of the warnings.)

Not all flags are supported by all modern compiler versions that are
still in use. Wrap all relatively new flags into 'supported' checks
which verify that the flag seems to be actually supported by the
compiler. Skip all flags thar are not supported: we're not going to
get any use from them anyway but this should not prevent compilation.

* Switch Soter and Themis makefiles

This has been triggering me for a while, but Soter build commands
belong to soter.mk and Themis build instruction should be located
in themis.mk.

This works now because both files are included into the top-level
makefile, but it is still quite surpising if you are looking to add
some Themis-specific changes to the test makefile.

* Make unused parameters as unused

Recently added "-Wunused-parameter" uncovered a bunch of unused
parameters. Mark genuinely unused parameters with UNUSED macro.
Most of these are "this" pointers in Java binding.

* Drop dead code and unused variables

Recently added "-Wunused-functions" and "-Wunused-variables" have found
some functions and variables which are really not used. Remove them.

* Fix signed-unsigned comparisons

Recently enabled "-Wall" warnings are now triggered by comparisons
between signed and unsigned values. Insert explicit casts to silence
these warnings. All affected comparisons are valid as signed values
are either checked for being negative, or should never be negative
in the first place (but legacy APIs are legacy).

* Correct const methods in ThemisPP

Recently enabled warnings were triggered by useless "const" specifiers
on scalar return values of some methods. These are indeed useless, but
the methods themselves are pure getters and may actually be marked
"const" instead.

* Strict function prototypes in C code

Recently added "-Wstrict-prototypes" has found a bunch of functions
(in unit tests) which do not have strict prototypes. In C functions
that have "()" argument lists and do not have an explicit prototype
are allowed to accept arbitrary amount of arguments.

Ensure that all nullary functions have a "(void)" argument list for
consistency with many other places which already use this spelling.

* Avoid variable-length arrays

VLAs have a history of being problematic from security perspective as
they are fairly easy to exploit for various buffer overflow and ROP
attacks. We have enabled "-Wvla" warning to weed them out. Replace
existing VLAs with explicit dynamic memory allocation.

* Silence a warning about incomplete initialization

Recently enabled "-Wmissing-field-initializers" (via "-Wextra") was
triggered by these struct initializers. Replace incomplete
initialization with an explicit default initialization. Since these
are static variables they are going to be zero-initialized anyway.

* Do not use uninitialized variables

This is an example of bug prevented by "-Wuninitialized" warning. Here
the "res" variable is not actually initialized during the test failure
handing. If it happens to be non-zero (which is quite likely) then the
failure of a test will not be reported.

secure_session_create() cannot possibly return a themis_status_t and
does not update "res". Just use "false" to always report a failure.

* Mark test_rand_with_nist() as unused for CI builds

This function is not actually called by CI due to -DCIRICLE_TEST flag.
(As NIST tests are often failing in Docker environment.) It looks like
dead code to compiler, but I don't want to hide it under an #ifdef so
just mark it as unused instead to simply silence the warning.

* Ensure full switch-case coverage

Prevent -Wswitch warnings by having an explicit default branch that
handles the error path.

* Don't build Rust integration tools quietly

Initially I thought that integration tool build should be successful
all the time so I have quieted the "cargo install" command. However,
it turned out that sometimes this build can fail and the error message
is very unhelpful.

Remove the "--quiet" flag and tweak the messages around it. Now the
build will be more verbose and will actually show error messages in
the build logs.

* Don't reexport CFLAGS and LDFLAGS

We don't want our CFLAGS and LDFLAGS to leak to child processes run
by Make (for example, to child Makes spawned by Cargo when building
other native libraries for Rust, or for the AFL-modified Themis which
is made by a recursively called Make). Unexport these two variables
which are important for C builds.

Exporting these variables now breaks Rust-Themis builds which have
some other native dependencies which produce warnings with our flags
which are now treated as build errors. We don't want Themis build flags
to affect any other libraries.

* Use special compiler flags for AFL-enabled builds

It turns out that afl-clang tends to have a completely different set
of supported flags than normal clang and gcc. Therefore we have to be
more picky about the flags we set in AFL_CFLAGS. Filter out all flags
that AFL_CC does not seem to support.

Teach "supported" function to accept an optional second argument
that sets the compiler to test. It will use $(CC) by default if
the argument is omitted.

* Do not treat warnings as errors by default

Enabling so many warnings is likely to be risky for any user builds and
may cause issues with forward compatibility (i.e., when a compiler adds
a new warning to -Wall set). Let's treat warnings as non-fatal by default
but fatal during CI builds. They users can also opt-in for warnings as
errors by setting WITH_FATAL_WARNINGS=yes.

* Use the latest compilers for analyzing the code

We need to set CC to the same compiler as used for formatting the code
in order for our "supported warning detector" to work accurately. Use
clang-7 that we use for formatting.