Use C constants directly
It turns out that CGo actually does export C defines as Go constants,
therefore we don't need to use intermediate variables for that.

Previous code actually does not compile with Go 1.9.2 that we have
on CI, but it is fine with my Go 1.12. The new code is fine in both
environments.