Deprecate incorrectly named API (#424)
* Rename cell.CELL_MODE_* constants
* Rename compare.COMPARE_* constants
* Rename keys.KEYTYPE_* constants
* Rename session.STATE_* constants

  - Provide new names that don't offend golint
  - Use new names throughout the code base
  - Mark the old names deprecated and redefine constants using new values

* Rename session.SecureSession.GetRemoteId method

  - Provide new name that don't offend golint
  - Use the new name throughout the code base
  - Mark the old method deprecated and reimplement it using the new one

* Use C constants directly

It turns out that CGo actually does export C defines as Go constants,
therefore we don't need to use intermediate variables for Secure
Comparator.

Previous code actually does not compile with Go 1.9.2 that we have
on CI, but it is fine with my Go 1.12. The new code is fine in both
environments.