Carthage code samples for iOS and macOS (#428)
This adds Xcode projects and code samples using Carthage packages
of Themis.

Since we already have CocoaPods examples, new ones made the directory
structure a bit messy. I have moved the code around so that it's a bit
easier to navigate:

    docs
    └── examples
        ├── objc
        │   ├── iOS-CocoaPods   (moved)
        │   ├── iOS-Carthage    (new)
        │   └── macOS-Carthage  (new)
        └── swift
            ├── iOS-CocoaPods   (moved)
            ├── iOS-Carthage    (new)
            └── macOS-Carthage  (new)

All samples use the same code, just copy-pasted around.

In order to initialize Carthage project one must run

    carthage bootstrap

in the project directory to pull the dependencies.

We pin Themis version to the current "master".

There is one caveat with new Objective-C samples: they import Themis as

    #import <themis/themis.h>

while CocoaPods samples use the following form:

    #import <objcthemis/objcthemis.h>

CocoaPods use some wild magic for that imports to be possible while
Carthage projects use the resulting framework name. I have not found
an easy way around that so it seems we'll have to live with this
discrepancy.