Retarget OpenSSL dependency
Use the upstream repository of OpenSSL instead of my private fork,
now that Marcin has merged Carthage support into the mainline.

Currently there is no suitable stable tag so pin a specific commit
instead that is known to work.

Update the lock-file as well so that we get correct versions after
"carthage bootstrap".