Use the latest stable version in examples
Remove version specifiers in Carthage examples which means that we'll
always use the latest stable (tagged) version. At the moment this is
provisional 0.10.5 version, then it will be 0.11, and so on.

Drop the lock files as well and add them into .gitignore. That way
the users will recreate the lockfiles as necessary and won't commit
them accidentally during development.