Update Carthage dependencies (#430)
* Retarget OpenSSL dependency

Use the upstream repository of OpenSSL instead of my private fork,
now that Marcin has merged Carthage support into the mainline.

Currently there is no suitable stable tag so pin a specific commit
instead that is known to work.

Update the lock-file as well so that we get correct versions after
"carthage bootstrap".

* Use the latest stable version in examples

Remove version specifiers in Carthage examples which means that we'll
always use the latest stable (tagged) version. At the moment this is
provisional 0.10.5 version, then it will be 0.11, and so on.

Drop the lock files as well and add them into .gitignore. That way
the users will recreate the lockfiles as necessary and won't commit
them accidentally during development.

* Revert "Use the latest stable version in examples"

This reverts commit d75a5a7055f70c14eb501a1dd41bdeb646c632a8.