Try standard system path when pkg-config fails
Currently Rust-Themis requires pkg-config to successfully find Themis.
If it doesn't (for whatever reason) then the crate won't compile.
However, the library might be correctly installed in the system under
standard paths. This allows us to easily link against it dynamically.
Let's support this use case to allow non-packaged installations.

If pkg-config successfully locates core Themis library then we just go
with it. Otherwise we effectively try to run

    cc -shared src/dummy.c -lthemis

and see if that compiles. If it does then apparently Themis is installed
and we can use it by emitting appropriate instructions to Cargo.

Note that we support only dynamic linkage in this case. Static linkage
will not work because we will have to figure out transitive dependencies
like Soter and OpenSSL, which is the reason for preferring pkg-config.