Relax compatibility constraints (#443)
Initially I thought that it would be good to pin "themis" to a
particular version of "libthemis-sys" package (more stability and
predictability, etc.) However, if you think about it, this strategy
requires updating "themis" crate whenever there is patch update in
"libthemis-sys" (ditto for "libthemis-sys" and "libthemis-src").
It may become a bother so I'd rather avoid it.

We adhere to semver, therefore patch version updates should be
compatible between minor versions. Relax the dependency constraints
to allow patch updates between the crates.