Try standard system path when pkg-config fails (#444)
Currently Rust-Themis requires pkg-config to successfully find Themis.
If it doesn't (for whatever reason) then the crate won't compile.
However, the library might be correctly installed in the system under
standard paths. This allows us to easily link against it dynamically.
Let's support this use case to allow non-packaged installations.

If pkg-config successfully locates core Themis library then we just go
with it. Otherwise we effectively try to run

    cc -shared src/dummy.c -lthemis

and see if that compiles. If it does then apparently Themis is installed
and we can use it by emitting appropriate instructions to Cargo.

Note that we support only dynamic linkage in this case. Static linkage
will not work because we will have to figure out transitive dependencies
like Soter and OpenSSL, which is the reason for preferring pkg-config.

* Wrap pkg_config::Library type

We're going to rely not only on pkg-config for locating Themis, but we
cannot construct pkg_config::Library instances (it has private fields).
Use a custom type with the same API instead.

* Don't panic immediately on pkg-config failure

We're going to provide an alternative code path for when pkg-config
fails, so instead of panicking with an error message just output it
to stderr and panic later. This enables us to insert some actions in
between and prevent the panic.

* Add changelog entry