Cleanup linkage: libraries
First of all, make sure that phony targets like "themis_shared" depend
on actual files in the build directory. This ensures that the libraries
are not relinked if their components did not actually change.

Then move library build rules into their respective submakefiles, and
clean them up somewhat. Now the rules are a bit more consistent, avoid
strange naming, and are located close to files that are collected into
resulting libraries.

While we're here, simplify linkage of the JNI wrapper library. Always
embed Themis, Soter, and cryptographic backend into the JNI wrapper so
that it actually depends only on OpenSSL if necessary. Never link JNI
wrapper against Themis library. Instead, libthemis_jni.so contains and
exports symbols of libthemis.so and libsoter.so.