Optimize embedded BoringSSL build (#447)
Don't build unnecessary libraries, binaries, tests, etc. Build only two
static libraries that we need. This results in considerable time saving
during the builds: BoringSSL contains quite a few tests and tools which
we do not run and use, yet they get built.

Native
------

Before:

    $ time make all ENGINE=boringssl

    real    4m36.341s
    user    3m44.060s
    sys     0m33.924s

After:

    $ time make all ENGINE=boringssl

    real    1m56.910s
    user    1m29.960s
    sys     0m19.392s

Nuff said.

Android
-------

Before:

    $ time ./gradlew --no-daemon --no-parallel --max-workers=2 assembleDebug

    real    34m46.028s
    user    29m9.580s
    sys     6m39.280s

After:

    $ time ./gradlew --no-daemon --no-parallel --max-workers=2 assembleDebug

    real    12m46.028s
    user    9m1.880s
    sys     2m52.020s

Admittedly, the difference may be smaller--as we still have to wait for
Gradle to download the whole Internet during Android build--but that's
still a significant improvement because there are four architectures
that we need to build BoringSSL for.