Cleanup linkage: tests
Do the same Makefile refactoring for test binaries. Move linkage rules
to submakefiles, update dependencies and make sure that tests are linked
against Themis statically (so that we don't use something installed in
the system, for example). Also pay attention to cryptographic backend,
we need to embed BoringSSL if we're building with it, or link against
external library otherwise.