Simplify static linkage
Introduce helper varibles that contain all libraries required for
linking against Soter and Themis statically. This keeps dependency lists
a bit shorter.

For Soter we need libsoter.a, and maybe libcrypto.a and libdecrepit.a if
BoringSSL is embedded.

For Themis we need libthemis.a and whatever is needed for Soter.