Avoid overlinking dynamic libraries
Currently we add cryptographic libraries to all targets by setting
LDFLAGS. This has an unwanted effect of linking libthemis.so against
libcrypto.so. We want libthemis.so to depend only on libsoter.so
(which in turn *may* depend on libcrypto.so, if we're not using
embedded BoringSSL).

Do not extend LDFLAGS when choosing cryptographic backend. Instead,
link each individual binary with CRYPTO_ENGINE_LDFLAGS as necessary.

Similarly, we have CRYPTO_ENGINE_CFLAGS that set header search path.
However, this path is required when building all libraries and tests
because all of them include private Soter headers that depend on the
selected cryptographic backend. It is easier to just extend CFLAGS.