Common Makefile rules: code formatting
Using consistent object file placement allows us to simplify formatting
build rules in the same way. All marker files will not be placed in the
same place as object files.

This allows us to remove a lot of special cases and use common
formatting rules for all C/C++ code.

clang-format and clang-tidy require correct compilation flags.
Thankfully, CFLAGS values are already correctly set everywhere so we do
not need to do anything special for that. Though, JsThemis is a bit
special, because we do not compile it normally.