Simplify package and library versioning
Current versioning approach in Makefile is way too complicated and does
not work for distributon tarballs when git repository is not available.
Fun fact: VERSION variable actually gets set twice.

Remove all this guesswork based on "git describe" and simply read the
version out of Makefile. Let it be the single source of truth.