Use 'make install' for packaging
Replace explicit 'file maps' for packaging with "make install"
invocations. This ensures that we always package the same files
that get installed when building from source.

The list of files to be packaged is still maintained separately as we
do split them into "libthemis" and "libthemis-dev" packages. Note that
the split is not entirely correct: the static libraries and symbolic
links should go into development package as well. However, for now we
keep the old split for compatibility.

We use the newly added DESTDIR capability to install into a temporary
staging directory and then collect the files from there. Now we also
no longer need to create symbolic links manually. However, stripping
binaries still has to be performed, and we still need to construct
pre- and postinstall scripts. Let's do it inline to save the trouble.