Detect and support Emscripten builds
The users are expected to build Themis for Wasm by using "emmake"
helper tool that sets up all necessary environment variables.
For example, CC variable will be set to Emscripten's cross-compiler.

Detect when we're building under emmake by inspecting the compiler type,
provide IS_EMSCRIPTEN variable if that's the case. We're going to use
this switch to customize a bunch of build steps.

First of all, we need to run CMake for embedded BoringSSL using the
"emconfigure" helper that sets its up properly. Additionally, we need
to disable assembly usage in BoringSSL's build as it detects the
platform incorrectly and tries using x86 assembly (which will not work
on Wasm targets, obviously).