Warning about multiple Themis installations
Changing default installation path from /usr to /usr/local is more or
less safe on Linux as the compilers and linker-loaders *usually* check
/usr/local before /usr. Therefore after "sudo make install" the software
will (probably) be using the new version of Themis. However, the old
installation may still remain there in "/usr" and this may cause issues
with non-standard configurations.

Check if there are multiple installations of Themis and print a warning
for the users, suggesting to keep only one installation in their system.