Improve Makefile rules for linkage (#450)
* Merge COVERLDFLAGS into LDFLAGS

Just use LDFLAGS everywhere. If we build Themis for coverage analysis
then we need to add relevant flags to all compiled libraries and tools.
There is no need to keep a separate variable for that.

* Cleanup linkage: libraries

First of all, make sure that phony targets like "themis_shared" depend
on actual files in the build directory. This ensures that the libraries
are not relinked if their components did not actually change.

Then move library build rules into their respective submakefiles, and
clean them up somewhat. Now the rules are a bit more consistent, avoid
strange naming, and are located close to files that are collected into
resulting libraries.

While we're here, simplify linkage of the JNI wrapper library. Always
embed Themis, Soter, and cryptographic backend into the JNI wrapper so
that it actually depends only on OpenSSL if necessary. Never link JNI
wrapper against Themis library. Instead, libthemis_jni.so contains and
exports symbols of libthemis.so and libsoter.so.

* Cleanup linkage: tests

Do the same Makefile refactoring for test binaries. Move linkage rules
to submakefiles, update dependencies and make sure that tests are linked
against Themis statically (so that we don't use something installed in
the system, for example). Also pay attention to cryptographic backend,
we need to embed BoringSSL if we're building with it, or link against
external library otherwise.

* Simplify static linkage

Introduce helper varibles that contain all libraries required for
linking against Soter and Themis statically. This keeps dependency lists
a bit shorter.

For Soter we need libsoter.a, and maybe libcrypto.a and libdecrepit.a if
BoringSSL is embedded.

For Themis we need libthemis.a and whatever is needed for Soter.

* Avoid overlinking dynamic libraries

Currently we add cryptographic libraries to all targets by setting
LDFLAGS. This has an unwanted effect of linking libthemis.so against
libcrypto.so. We want libthemis.so to depend only on libsoter.so
(which in turn *may* depend on libcrypto.so, if we're not using
embedded BoringSSL).

Do not extend LDFLAGS when choosing cryptographic backend. Instead,
link each individual binary with CRYPTO_ENGINE_LDFLAGS as necessary.

Similarly, we have CRYPTO_ENGINE_CFLAGS that set header search path.
However, this path is required when building all libraries and tests
because all of them include private Soter headers that depend on the
selected cryptographic backend. It is easier to just extend CFLAGS.