Support Emscripten in Soter and Themis tests
We can build Soter and Themis tests and run them, but there are a couple
of caveats here.

First of all, emcc output depends on file extension. File names
"themis_test" and "soter_test" result in LLVM bitcode being packed in
them, which is not what we want. We should use "*.js" extension to get
something runnable. Introduce intermediate variables to rename the tests
based on the target environment. Use "themis_test.js" for Emscripten and
old "themis_test" for everything else.

Additionally, we use SINGLE_FILE flag to produce a single *.js file
instead of a separate *.wasm file with compiled code and a *.js loader
for it. This makes it easier to launch the tests as we don't have to
change directories and don't generate extra files.

It also turned out that popen() function is not available in Emscripten
runtime. Because of that we cannot use NIST STS as is. Disable it for
now. Hopefully, we'll figure out how to use it later.

The next point is that we cannot run resulting tests directly. We should
use Node.js to interpret then, not execute *.js files (which is usually
not possible on typical systems). Therefore, check if "node" binary is
available and use it to run tests when requested. But only in Emscripten
environment. If Node.js is not available then print a warning and fail
the build.